Move compare.go to filter
This still needs a lot more cleanup (specifically consolidation with
the similar logic that exists for expressions) but this is only used
from filter/ so in preparation for repurposing zx to be about native
zng values, put it with filter for now.
